Make the STYLE_BLOOM global private.
